Adds background to signup
