+++ src/App.js
@@ -6,17 +6,17 @@ import BrowseList from './pages/BrowseList';
 import Landing from './pages/Landing';
 import Navbar from './components/Navbar';
 import AuthRoute from './components/AuthRoute';
-import DefaultLayout, { HomePage } from './components/Layout';
+import DefaultLayout, { WithBackground } from './components/Layout';
 
 function App() {
   return (
     <Fragment>
       <Navbar />
       <Switch>
-        <HomePage exact path="/" component={Landing} />
+        <WithBackground exact path="/" component={Landing} />
         <AuthRoute exact path="/browse" component={BrowseList} />
         <DefaultLayout exact path="/login" component={Login} />
-        <DefaultLayout exact path="/signup" component={Signup} />
+        <WithBackground exact path="/signup" component={Signup} />
       </Switch>
     </Fragment>
   );
